essage=initial rendering
