+++ src/js/modules/recruitment/components/RecruitmentPage.js
@@ -1,14 +1,29 @@
 import React, { PureComponent } from "react";
 import injectSheet from "react-jss";
 
-import { Section } from "./Section";
 import * as constants from "../constants";
 
 const styles = {
-    
-}
+    quote: {
+        background: {
+            borderRadius: "15px",
+            borderColor: "gray",
+            backgroundColor: "gainsboro",
+        },
+        text: {
+            fontFamily: "Minion Pro",
+            fontSize: "15px"
+        },
+    },
+    section: {
 
-const RecruitmentPage = () => {
+    },
+    page: {
+
+    }
+};
+
+const RecruitmentPage = ({}) => {
     //const sections = constants.SECTIONS;
     return (
         /*<div>
@@ -18,7 +33,31 @@ const RecruitmentPage = () => {
         </div>
         */
        <div>
-           <Section name="Web" description="Best department" quotes={{text: "Fight the power", source: "Nich"}} />
+           <Section name="Web" description="Best department" quotes={[{text: "Fight the power", source: "Nich"}]} />
+        </div>
+    );
+};
+
+
+const Quote = ({ text, source }) => {
+    return (
+        <div id={styles.quote.background}>
+            <p id={styles.quote.text}>
+                {"“" + text + "”"}
+            </p>
+            <i id={styles.quote.text}>{"―" + source}</i>
+        </div>
+    );
+};
+
+const Section = ({ name, description, quotes }) => {
+    return (
+        <div>
+            <h1>{name}</h1>
+            <p>{description}</p>
+            {quotes.map(({text, source}) => {
+                return (<Quote text={text} source={source} />)})
+            }
         </div>
     );
 };
