--- src/js/modules/recruitment/components/Section.js
@@ -1,25 +1,0 @@
-import React from "react";
-import injectSheet from "react-jss";
-import { Quote } from "./Quote";
-
-const styles = {
-    background: {
-        backgroundColor: "powderblue",
-        borderRadius: "15px",
-        borderColor: "steelblue"
-    }
-}
-
-const Section = ({ name, description, quotes }) => {
-    return (
-        <div>
-            <h1>{name}</h1>
-            <p>{description}</p>
-            {quotes.map((text, source) => {
-                return (<Quote text={text} source={source} />)})
-            }
-        </div>
-    );
-};
-
-export default injectSheet(styles)(Section);
