+++ src/js/modules/recruitment/components/index.js
@@ -1,3 +1,1 @@
 export {default as RecruitmentPage} from "./RecruitmentPage";
-export {default as Section} from "./Section";
-export {default as Quote} from "./Quote";
