+++ src/js/versionInfo.js
@@ -3,5 +3,5 @@
  *           without modifying formatting for automation scripts
  * @type {{VERSION: string}}
  */
-let VersionInfo = { VERSION : '2.0.31' };
+let VersionInfo = { VERSION : '2.0.51' };
 export default VersionInfo;
